requirements.txt
asn1crypto==0.24.0
asym-crypto-yaml==0.0.5
backcall==0.1.0
backoff==1.8.0
boto3==1.9.116
botocore==1.12.116
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
decorator==4.3.0
docutils==0.14
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
jmespath==0.9.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
